Refactor: 상세 임시데이터 분리 #103
