+++ lib/features/place/presentation/models/place_detail_role.dart
@@ -1,0 +1,9 @@
+enum PlaceDetailRole { leader, member }
+
+PlaceDetailRole? placeDetailRoleFromQuery(String? value) {
+  return switch (value) {
+    'leader' => PlaceDetailRole.leader,
+    'member' || 'team' => PlaceDetailRole.member,
+    _ => null,
+  };
+}
